Controller data validation added

+++ core/src/main/java/com/vladimir/mailserver/controller/AddressRestController.java
@@ -2,6 +2,7 @@ package com.vladimir.mailserver.controller;
 
 import com.vladimir.mailserver.dto.AddressDto;
 import com.vladimir.mailserver.service.AddressService;
+import com.vladimir.mailserver.service.ValidatorService;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.security.core.context.SecurityContextHolder;
@@ -14,9 +15,11 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/address")
 public class AddressRestController {
     private AddressService addressService;
+    private ValidatorService validatorService;
 
-    public AddressRestController(AddressService addressService) {
+    public AddressRestController(AddressService addressService, ValidatorService validatorService) {
         this.addressService = addressService;
+        this.validatorService = validatorService;
     }
 
     @GetMapping("/get")
@@ -27,19 +30,28 @@ public class AddressRestController {
 
     @GetMapping("/add/{name}/{address}")
     public boolean addAddress(@PathVariable String name, @PathVariable String address) {
-        String login = SecurityContextHolder.getContext().getAuthentication().getName();
-        return addressService.addAddress(login, name, address);
+        if (validatorService.validateAddressName(name) && validatorService.validateEmailAddress(address)) {
+            String login = SecurityContextHolder.getContext().getAuthentication().getName();
+            return addressService.addAddress(login, name, address);
+        }
+        return false;
     }
 
     @GetMapping("/delete/{id}")
     public boolean deleteContact(@PathVariable Long id) {
-        String login = SecurityContextHolder.getContext().getAuthentication().getName();
-        return addressService.deleteAddress(login, id);
+        if (id != null && id > 0) {
+            String login = SecurityContextHolder.getContext().getAuthentication().getName();
+            return addressService.deleteAddress(login, id);
+        }
+        return false;
     }
 
     @GetMapping("/edit/{id}/{name}/{address}")
     public boolean editContact(@PathVariable Long id, @PathVariable String name, @PathVariable String address) {
-        String login = SecurityContextHolder.getContext().getAuthentication().getName();
-        return addressService.editAddress(login, id, name, address);
+        if (id != null && id > 0 && validatorService.validateAddressName(name) && validatorService.validateEmailAddress(address)) {
+            String login = SecurityContextHolder.getContext().getAuthentication().getName();
+            return addressService.editAddress(login, id, name, address);
+        }
+        return false;
     }
 }
